# src/api-docs/swagger.yaml
openapi: 3.1.0
...
info:
    title: Sentinel BankServer API
    version: 1.0.0
    summary: Encrypted-at-the-application-layer banking API.
    description: |
        Reference documentation for the BankServer HTTP surface. Every operation
        is grouped under one of the 18 numbered flow tags so the call order is
        self-evident. Each operation also carries:

          * `description` — markdown including a bulleted **Prerequisites** list
            of operationIds that must run before this one.
          * `x-call-order` — an integer that orders operations *within* the flow
            tag. Tools that respect extension fields can sort by this.
          * `x-depends-on` — machine-readable mirror of the prerequisites list.
          * `x-decrypted-request` / `x-decrypted-response` — for encrypted ops,
            the JSON shape carried inside the inner envelope's `data` field
            (i.e. the operation's actual payload). The wire-level
            `requestBody` / `responses` schemas keep the literal wire envelope
            (`{ sessionId, payload }` in / `{ payload }` out); see
            **Cryptographic envelope** below for the inner envelope.

        ## Cryptographic envelope

        After the bootstrap handshake (flow `01-Bootstrap`), every encrypted
        endpoint uses **two nested envelopes**:

        **1. Wire envelope** — what travels over HTTP:

        ```json
        // Request
...

        // Response
...
        ```

...

...

        ```json
...
        ```

...

...

...

...

...

...

...
servers:
...
security:
    - SessionId: []
...
components:
...
    securitySchemes:
        SessionId:
            type: apiKey
            in: header
            name: x-session-id
            description: |
                UUID identifying the encrypted session created by the bootstrap
                handshake. Required on every encrypted endpoint.
        ActionToken:
            type: apiKey
            in: header
            name: x-action-token
            description: |
                One-shot token minted by `postWebauthnAuthenticationVerify` and
                bound to a specific `(action, paramsHash, sessionId)` tuple.
                Required by every step-up gated operation.
        OtpToken:
            type: apiKey
            in: header
            name: x-otp-token
            description: |
                One-shot token minted by `postIdentityOtpVerify` and bound
                to a specific `(action, paramsHash, sessionId)` tuple.
                Required ALONGSIDE `ActionToken` by every OTP-gated route
                (passkey.add, passkey.revoke, password.change, session.wipe,
                admin.recovery).
    schemas:
...
        ApiError:
            type: object
            required: [message]
            properties:
                message:
                    type: string
                requestId:
                    type: string
                details:
                    type: object
        ApiErrorEnvelope:
            type: object
            description: |
                Decrypted error body. On encrypted routes this whole object
                rides inside the
                [`DecryptedEnvelope`](#/components/schemas/DecryptedEnvelope)'s
                `data` field (which is itself wrapped in
                `EnvelopedResponse.payload` on the wire). On plaintext
                bootstrap routes (`/security/*`) it comes back as the raw
                response body.
            required: [success, error]
            properties:
                success:
                    type: boolean
                    const: false
                error:
                    $ref: "#/components/schemas/ApiError"

        # ── Identity ───────────────────────────────────────────────────────────
...
        SuccessResult:
            type: object
            required: [success]
            properties:
                success:
                    type: boolean
                    const: true

        # ── Audit ──────────────────────────────────────────────────────────────
...
        SecuritySessionResponse:
            type: object
            required: [sessionId, serverPublicJwk, salt]
            properties:
                sessionId:
                    type: string
                    format: uuid
                serverPublicJwk:
                    type: object
                salt:
                    type: string
        SecurityHandshakeRequest:
            type: object
            required: [sessionId, clientPublicJwk]
            properties:
                sessionId:
                    type: string
                    format: uuid
                clientPublicJwk:
                    type: object

        # ── Dev ────────────────────────────────────────────────────────────────
...
paths:
...
    /security/session:
        get:
            operationId: getSecuritySession
            tags: [01-Bootstrap]
            summary: Allocate a session and return the server's ECDH public key.
            description: |
                First step of the encryption handshake. Plaintext.
                **Prerequisites:** none.
            security: []
            x-call-order: 1
            x-depends-on: []
            responses:
                "200":
                    description: New session allocated.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SecuritySessionResponse"
                "4XX":
                    description: Validation error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApiErrorEnvelope"
    /security/handshake:
        post:
            operationId: postSecurityHandshake
            tags: [01-Bootstrap]
            summary: Complete the ECDH handshake by submitting the client public key.
            description: |
                Derives the AES session key. Following endpoints are encrypted.
                **Prerequisites:** `getSecuritySession`.
            security: []
            x-call-order: 2
            x-depends-on: [getSecuritySession]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/SecurityHandshakeRequest"
            responses:
                "200":
                    description: Key live on both sides.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SuccessResult"
                "4XX":
                    description: Handshake error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApiErrorEnvelope"

    # =========================================================================
    # 02-Identity-Auth
    # =========================================================================
...